This modification contains 157 changed regions. <ol>
<li> Moved a lot of common functionality into libjconfigure.bash, a
new file in this directory (this could be moved).  libjconfigure.bash
includes an 'err' function.  This consolidates two different versions
of the error handling that had been in place, one for jconfigure and
the other for the jbuild* files it created.  The jbuild* files each
contain an inline copy of libjconfigure.bash.
<p>
Since we set up the error handling in libjconfigure.bash, it means
that every jbuild* file now has the exit-on-error checking enabled.
<p>
Everywhere that we copy files we now use copyIfNewer.

<li>__linuxsmp__ has now been officially exorcised from the source.

<li>Wherever we invoked "cd" and might want to tell the user or his
Emacs process that we've changed directories,, we now invoke the
chdir() function.  This gives us better error-finding for Emacs, when
we're running with --trace=make.

<li>As usual, I wrote more comments.
<li>Minor format improvements, where it seemed appropriate.

<li>We now consistently apply PREPROC_FLAGS everywhere we invoke the preprocessor.

<li>We now make liberal use of the "run()" function, also in
libjconfigure.bash.  If you build with --trace, you actually get
all of the relevant commands printed out; wow!

<li>We now run jconfigure and all jbuild* generated programs
with the shell option 'noundef'; this means that we
complain about any undefined variables.  To make this work, we now
explicitly define an RVM_WITH_<i>GC_TYPE</i> variable for each of the GC
types.  We have to be careful about adding configuration variables
now; in cases where it's ok for the empty string to replace an unset
variable, refer to that variable as: ${POSSIBLY_UNSET_VARIABLE-}

<li>
Rewrote some of the error message display code.  We now put
the long messages through a new function, show_mesg_raw().

<li> We got rid of the warning we used to display about not having a
shell that knows about the ERR Trap.

<li> The sed expression for cleaning lists is now encapsulated into
the cleanFileList() function.

<li> Replaced a couple of explicit references to \$JAL_ROOT with ones
to $ROOT_DIRECTORY_STR.  Replaced explicit references to /bin/bash
with references to ${BASH}.  In general, found places where we weren't
modularizing explicit file names properly and modularized them.  Use $CLASSES_DIRECTORY_STR where appropriate.

<li>Added yet more quoting, as usual.

<li>Added more cleanup handling so we properly delete PRIMORDIALS.1 etc. even
if we abort.  We also now delete the target of the files we copy if
we get interrupted or encounter trouble.
<p>
jconfigure now uses the SED variable to invoke sed ... hey,
if we're not going to set it, why use it?  There's also a SED_STR, etc.

<li> Better process for copying preprocessModifiedFiles to
jbuild.prep.host and jbuild.prep.target.

<li>
Only copy classpath.jar to rvmrt.jar if it needs copying.

<li> Added code to disable certain compiler -W flags.

<p>We do this by getting rid of the little functions run_booter_cxx()
and run_booter_cc().  Instead, we went back to setting BOOTER_CC, etc,
to the actual text of the commands to run.  Then we perform string
substitution on that text.  For example, when compiling a source file
where we don't want to get the -Wmissing-noreturn warning, we yank out
-Wmissing-noreturn from the string.

<li> File loading in "jconfigure" now encapsulated into load().

<li>Generate a new script, jbuild.install, that actually lets us
install jikes RVM into another directory.  Handy way to save off a
configuration that works before making changes.  Also lets us make a
binary distribution of Jikes RVM, shoujld it be necessary.

<li>No longer warn about plugin version #s being too long for Eclipse
until we actually try to install Eclipse.

</ol>