1) Improved coments  <br> 2) Set ME so we can test this independently of the main jconfigure program. <br> 3) chdir() now makes use of $PWD.  This is better for relative paths, not that we have any calls to chdir that use them, but still wise. <br>
4) Rewrote error handling yet again.  This in order to handle a change
in bash versions after Bash 2.05a: line numbers are now always
relative to the script, rather than relative to the line in a
function.   We now print a clear message.  We also test for what the
case is, using the new function set_have_function_relative_LINENO().
<ol>
<li>We now give out a much better message for shells that don't
implement the ERR trap.

<li> We handle the case of getting negative line #s in LINENO, such as
happens when we receive a signal.  The message we print is actually
nice now.

<li> Got rid of bogus signal-number printing code in the signal
handler; there doesn't seem to be a way to get at that data.

</ol>