Further modification to LD_LIBRARY_PATH setting in environment.host and environment.bash, so that the user's LD_LIBRARY_PATH doesn't override the HOST_JAVA_LD_LIBRARY_PATH, and so that the code is less messy.
In the very unlikely event that HOST_JAVA_LD_LIBRARY_PATH was unset,
but LD_LIBRARY_PATH were set, we would have faced a single trailing
colon being added to LD_LIBRARY_PATH, which might on some systems mean
that the current working directory would be searched before the system
libraries.  This wouldn't have affected the Jikes RVM build, since we
have to "cd" to RVM_BUILD before we call jbuild, but it's cleaner to fix
this anyway.