rvm/src/vm/compilers/baseline/GCMap VM_ReferenceMaps.java,1.42,1.43 Update of /cvsroot/jikesrvm/rvm/src/vm/compilers/baseline/GCMap In directory sc8-pr-cvs6.sourceforge.net:/tmp/cvs-serv11506/rvm/src/vm/compilers/baseline/GCMap
Modified Files:
	VM_ReferenceMaps.java
Log Message:
Fix to baseline GC maps.  Kudos to Daniel for this.

The problem is that when jsr maps are merged with the parent normal map, the
use of a slot as a code reference should dominate: the entry should not show
up in both code and object reference maps.  This fix excludes an entry from
the mereged reference map if it is a code reference.