Fixes to get Classpath CVS head working again (NB there's a bug in Classpath's jcl.c that needs patching from their end).
Moved support for JNI_OnLoad into a native method so that JNI calls work
as expected from JNI_OnLoad methods.

Modifications to the boot image writer and VM.boot so that certain known
fields are populated during boot image write time, thereby avoiding the
need to run class initializers in VM.boot. Typically we fail to populate
a field value when it's part of the bootstrap JVM and in some way
protected.

The patch also modifies how we count nullified references in the boot
image as previously values that were genuinely null would also be
counted as having been nullified by the boot image writer.