Fix to what appears to be a long standing stack-scanning bug.
Since revision 3170, trap frames have been skipped during stack scanning. This is apparently a bug because a side effect of calling getNextReferenceAddress() is that registeres saved by the callee are put into registerLocations data structure. 

This means that when a GC occurs with a trap frame on the stack, it will not handle the trapping method propoerly.