Delete obsolete SynchronizedObject annotation.
This was a leftover from the ECOOP'02 one word object model.  Explicit annotations like this actually aren't that useful since it turns out that in the Java 5 libraries there are a number of common coding patterns (eg Synchronized Collections) that violate the assumptions in the simple heuristics we had implemented.  If we ever re-visit this, we'll need to implement a better way to figure out which classes without synchronized virtual methods should be given inline thinLocks.