Delete unused Classpath-.97-2.txt
Add CPStringBuilder and a few other gnu.java.util.regex to the 
primordials so that we avoid compiling unresolved references into 
hot methods in the bootimage.

Adding CPStringBuilder to the bootimage enables scalar replacement to get rid 
of the CPStringBuilder instances in the key regex methods used by luindex.
This helps some, but there are still performance differences due to
3x space used in char[] between 0.97 and 0.98.  My current assumption is 
that this is due to different default sizing between CPStringBuilder 
and StringBuilder....continuing to investigate; not ready to merge yet.