RVM-814 : Inconsistent wrapping of Exceptions in InvocationTargetException
Fix underlying bug instead of compensating for symptoms. I believe the root cause of the problem is that the bytecode-generating reflection implementation did not match the exception handling behavior of the original out-of-line invoke implementation. These two implementations have to have the same exception behavior, or there is no way for Reflection.invoke to be correct as it conditionally calls one or the other depending on various configuration parameters.